+++ src/apps/enzyme-finance/ethereum/enzyme-finance.vault.token-fetcher.ts
@@ -1,7 +1,7 @@
 import { Inject } from '@nestjs/common';
 import BigNumber from 'bignumber.js';
 import { gql } from 'graphql-request';
-import _ from 'lodash';
+import _, { compact } from 'lodash';
 
 import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { Register } from '~app-toolkit/decorators';
@@ -73,16 +73,22 @@ export class EthereumEnzymeFinanceVaultTokenFetcher extends AppTokenTemplatePosi
     return (await contract.getTrackedAssets()).map(x => x.toLowerCase());
   }
 
-  async getPrice({ appToken, contract }: PriceStageParams<EnzymeFinanceVault, DefaultDataProps>): Promise<number> {
+  async getPrice({
+    appToken,
+    contract,
+    tokenLoader,
+  }: PriceStageParams<EnzymeFinanceVault, DefaultDataProps>): Promise<number> {
     const multicall = this.appToolkit.getMulticall(this.network);
-    const tokenSelector = this.appToolkit.getBaseTokenPriceSelector({ tags: { network, appId } });
-    const baseTokens = await tokenSelector.getAll({ network });
 
     const decimalsRaw = await this.getDecimals(contract);
     const supplyRaw = await this.getSupply(contract);
     const decimals = new BigNumber(10).exponentiatedBy(decimalsRaw);
     const supply = new BigNumber(supplyRaw.toString()).div(decimals);
     const underlying = await this.getUnderlyingTokenAddresses(contract);
+
+    const tokenDependencies = await tokenLoader
+      .getMany(underlying.map(tokenAddressRaw => ({ network, address: tokenAddressRaw.toLowerCase() })))
+      .then(deps => compact(deps));
 
     const totalAssetUnderManagement = _.sum(
       await Promise.all(
@@ -95,7 +101,7 @@ export class EthereumEnzymeFinanceVaultTokenFetcher extends AppTokenTemplatePosi
           ]);
 
           const amount = Number(tokenAmountRaw) / 10 ** decimals;
-          const baseToken = baseTokens.find(v => v.address === tokenAddress);
+          const baseToken = tokenDependencies.find(v => v.address === tokenAddress);
           if (!baseToken) return 0;
 
           return baseToken.price * amount;