+++ src/apps/pickle/ethereum/pickle.jar.token-fetcher.ts
@@ -14,11 +14,11 @@ import { PickleContractFactory, PickleJar } from '../contracts';
 import { PickleApiJarRegistry } from '../helpers/pickle.api.jar-registry';
 import { PICKLE_DEFINITION } from '../pickle.definition';
 
-@Register.TokenPositionFetcher({
-  appId: PICKLE_DEFINITION.id,
-  groupId: PICKLE_DEFINITION.groups.jar.id,
-  network: Network.ETHEREUM_MAINNET,
-})
+const appId = PICKLE_DEFINITION.id;
+const groupId = PICKLE_DEFINITION.groups.jar.id;
+const network = Network.ETHEREUM_MAINNET;
+
+@Register.TokenPositionFetcher({ appId, groupId, network })
 export class EthereumPickleJarTokenFetcher implements PositionFetcher<AppTokenPosition> {
   constructor(
     @Inject(PickleContractFactory)
@@ -28,7 +28,6 @@ export class EthereumPickleJarTokenFetcher implements PositionFetcher<AppTokenPo
   ) {}
 
   async getPositions() {
-    const network = Network.ETHEREUM_MAINNET;
     const vaults = await this.jarRegistry.getJarDefinitions({ network });
 
     return this.appToolkit.helpers.vaultTokenHelper.getTokens<PickleJar>({