+++ CHANGELOG.md
@@ -4,6 +4,56 @@
 
 [1]: https://www.npmjs.com/package/@zapper-fi/studio?activeTab=versions
 
+## [0.170.0](https://github.com/Zapper-fi/studio/compare/v0.169.0...v0.170.0) (2022-08-11)
+
+
+### Features
+
+* **template:** Add token loader as stage param ([#1146](https://github.com/Zapper-fi/studio/issues/1146)) ([70ef238](https://github.com/Zapper-fi/studio/commit/70ef23859d0d740f97cdce7c2f5f1189674fefc0))
+
+## [0.169.0](https://github.com/Zapper-fi/studio/compare/v0.168.0...v0.169.0) (2022-08-11)
+
+
+### Features
+
+* **token-deps:** Exclusively use `getTokenDependencySelector` as a selector ([#1144](https://github.com/Zapper-fi/studio/issues/1144)) ([828d69b](https://github.com/Zapper-fi/studio/commit/828d69b8f9905f40fe74373a18b66731763bff1b))
+
+## [0.168.0](https://github.com/Zapper-fi/studio/compare/v0.167.0...v0.168.0) (2022-08-11)
+
+
+### Features
+
+* **template:** Use token dependency resolution ([#1142](https://github.com/Zapper-fi/studio/issues/1142)) ([42d7c90](https://github.com/Zapper-fi/studio/commit/42d7c908f69177a13d0e59894db6622d43b4c322))
+
+## [0.167.0](https://github.com/Zapper-fi/studio/compare/v0.166.0...v0.167.0) (2022-08-11)
+
+
+### Features
+
+* **app-toolkit:** Add support for token dependency selection ([#1140](https://github.com/Zapper-fi/studio/issues/1140)) ([f278b1f](https://github.com/Zapper-fi/studio/commit/f278b1fc881aa4ce859838a24464a0b466853bcd))
+
+## [0.166.0](https://github.com/Zapper-fi/studio/compare/v0.165.0...v0.166.0) (2022-08-11)
+
+
+### Features
+
+* **multicall:** Return error do not throw ([00ecfda](https://github.com/Zapper-fi/studio/commit/00ecfdaee6a88d4566a9930222c20939c6a63ceb))
+
+## [0.165.0](https://github.com/Zapper-fi/studio/compare/v0.164.0...v0.165.0) (2022-08-11)
+
+
+### Features
+
+* **plutus:** Fix labels ([f1a9d9d](https://github.com/Zapper-fi/studio/commit/f1a9d9d2be30199a5fcac038ce1d4012fbc6528e))
+* **plutus:** Refactor Plutus to use templates and add new farms ([#1132](https://github.com/Zapper-fi/studio/issues/1132)) ([9bc6c0a](https://github.com/Zapper-fi/studio/commit/9bc6c0ac241308222b74a46f5f5dd8d7d0d37005))
+* **rubicon:** Clean-up unused contracts ([dfab021](https://github.com/Zapper-fi/studio/commit/dfab021f6aef61f89b96f696f5988a4f4d28202f))
+
+
+### Bug Fixes
+
+* **pickle:** Missing jars on polygon ([#1130](https://github.com/Zapper-fi/studio/issues/1130)) ([646d3da](https://github.com/Zapper-fi/studio/commit/646d3da3bbddc38001475363cd56473899d26ebf))
+* **rubicon:** Fix decimal issues on non-18 decimal coins ([#1121](https://github.com/Zapper-fi/studio/issues/1121)) ([f7a4896](https://github.com/Zapper-fi/studio/commit/f7a4896c246b73b5c4d2c61876fb0186e5bc5e8f))
+
 ## [0.164.0](https://github.com/Zapper-fi/studio/compare/v0.163.0...v0.164.0) (2022-08-11)